requirements.txt
psycopg2-binary
sqlalchemy
streamlit